Kamthe/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
# Note: Add FIRECRAWL_API_KEY to your .env file for firecrawl_scraper.py
